# openapi.json
{"openapi":"3.1.0","info":{"title":"Sistema Imbatible","description":"RPG de la vida real para optimizar comportamiento y productividad","version":"1.0.0"},"paths":{"/api/usuarios":{"post":{"summary":"Crear Usuario","operationId":"crear_usuario_api_usuarios_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsuarioCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsuarioResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/usuarios/{usuario_id}":{"get":{"summary":"Obtener Usuario","operationId":"obtener_usuario_api_usuarios__usuario_id__get","parameters":[{"name":"usuario_id","in":"path","required":true,"schema":{"type":"integer","title":"Usuario Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsuarioResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/usuarios/{usuario_id}/desactivar":{"patch":{"summary":"Desactivar Usuario","operationId":"desactivar_usuario_api_usuarios__usuario_id__desactivar_patch","parameters":[{"name":"usuario_id","in":"path","required":true,"schema":{"type":"integer","title":"Usuario Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/usuarios/{usuario_id}/actividades":{"post":{"summary":"Registrar Actividad","operationId":"registrar_actividad_api_usuarios__usuario_id__actividades_post","parameters":[{"name":"usuario_id","in":"path","required":true,"schema":{"type":"integer","title":"Usuario Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActividadRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActividadResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"summary":"Obtener Actividades","operationId":"obtener_actividades_api_usuarios__usuario_id__actividades_get","parameters":[{"name":"usuario_id","in":"path","required":true,"schema":{"type":"integer","title":"Usuario Id"}},{"name":"dias","in":"query","required":false,"schema":{"type":"integer","default":7,"title":"Dias"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/usuarios/{usuario_id}/hitos":{"post":{"summary":"Registrar Hito","operationId":"registrar_hito_api_usuarios__usuario_id__hitos_post","parameters":[{"name":"usuario_id","in":"path","required":true,"schema":{"type":"integer","title":"Usuario Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HitoRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HitoResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"summary":"Obtener Hitos","operationId":"obtener_hitos_api_usuarios__usuario_id__hitos_get","parameters":[{"name":"usuario_id","in":"path","required":true,"schema":{"type":"integer","title":"Usuario Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/usuarios/{usuario_id}/estadisticas":{"get":{"summary":"Estadisticas","operationId":"estadisticas_api_usuarios__usuario_id__estadisticas_get","parameters":[{"name":"usuario_id","in":"path","required":true,"schema":{"type":"integer","title":"Usuario Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"ActividadRequest":{"properties":{"tipo":{"$ref":"#/components/schemas/TipoActividad"},"duracion_minutos":{"type":"integer","title":"Duracion Minutos"},"timestamp":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Timestamp"},"descripcion":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Descripcion"}},"type":"object","required":["tipo","duracion_minutos"],"title":"ActividadRequest"},"ActividadResponse":{"properties":{"id":{"type":"integer","title":"Id"},"usuario_id":{"type":"integer","title":"Usuario Id"},"tipo":{"$ref":"#/components/schemas/TipoActividad"},"duracion_minutos":{"type":"integer","title":"Duracion Minutos"},"timestamp":{"type":"string","format":"date-time","title":"Timestamp"},"xp_generado":{"type":"integer","title":"Xp Generado"}},"type":"object","required":["id","usuario_id","tipo","duracion_minutos","timestamp","xp_generado"],"title":"ActividadResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HitoRequest":{"properties":{"tipo":{"$ref":"#/components/schemas/HitoTipo"},"valor_xp":{"type":"integer","title":"Valor Xp"},"descripcion":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Descripcion"},"timestamp":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Timestamp"}},"type":"object","required":["tipo","valor_xp"],"title":"HitoRequest"},"HitoResponse":{"properties":{"id":{"type":"integer","title":"Id"},"usuario_id":{"type":"integer","title":"Usuario Id"},"tipo":{"$ref":"#/components/schemas/HitoTipo"},"valor_xp":{"type":"integer","title":"Valor Xp"},"descripcion":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Descripcion"},"timestamp":{"type":"string","format":"date-time","title":"Timestamp"}},"type":"object","required":["id","usuario_id","tipo","valor_xp","descripcion","timestamp"],"title":"HitoResponse"},"HitoTipo":{"type":"string","enum":["energia","disciplina","enfoque","penalizacion"],"title":"HitoTipo"},"Nivel":{"type":"integer","enum":[1,2,3,4,5,6],"title":"Nivel"},"TipoActividad":{"type":"string","enum":["sueño","ejercicio","estudio","trabajo","transporte","musica"],"title":"TipoActividad"},"UsuarioCreate":{"properties":{"nombre":{"type":"string","title":"Nombre"},"email":{"type":"string","title":"Email"}},"type":"object","required":["nombre","email"],"title":"UsuarioCreate"},"UsuarioResponse":{"properties":{"id":{"type":"integer","title":"Id"},"nombre":{"type":"string","title":"Nombre"},"email":{"type":"string","title":"Email"},"xp_total":{"type":"integer","title":"Xp Total"},"nivel":{"$ref":"#/components/schemas/Nivel"},"racha_dias":{"type":"integer","title":"Racha Dias"},"última_actividad":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Última Actividad"}},"type":"object","required":["id","nombre","email","xp_total","nivel","racha_dias","última_actividad"],"title":"UsuarioResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}